requirements.txt
supabase
psycopg2-binary
python-dotenv
discord.py
chat-exporter
groq
requests
qrcode